revert

+++ src/main/java/com/example/mainproject/controller/ActionController.java
@@ -24,7 +24,7 @@ public class ActionController {
     @PostMapping
     public ResponseEntity createAction(@RequestBody ActionEntity actionEntity,
                                        @RequestParam Integer well_id,
-                                       @RequestParam Integer action_type_id) {
+                                       @RequestParam Long action_type_id) {
         try {
             return ResponseEntity.ok(actionService.createData(actionEntity, well_id, action_type_id));
         } catch (Exception e) {
