+++ src/main/java/com/example/mainproject/controller/HangerTypeController.java
@@ -33,7 +33,7 @@ public class HangerTypeController {
     }
 
     @GetMapping
-    public ResponseEntity getDeptType(@RequestParam Integer hanger_type_id) {
+    public ResponseEntity getDeptType(@RequestParam Long hanger_type_id) {
         try {
             return ResponseEntity.ok(hangerTypeService.getOne(hanger_type_id));
         } catch (Exception e) {
@@ -42,7 +42,7 @@ public class HangerTypeController {
     }
 
     @DeleteMapping("/{hanger_type_id}")
-    public ResponseEntity deleteDeptType(@PathVariable Integer hanger_type_id) {
+    public ResponseEntity deleteDeptType(@PathVariable Long hanger_type_id) {
         try {
             return ResponseEntity.ok("Тип подвески удален: " + hangerTypeService.deleteOne(hanger_type_id));
         } catch (Exception e) {
@@ -51,7 +51,7 @@ public class HangerTypeController {
     }
 
     @PutMapping
-    public ResponseEntity updateHangerType(@RequestParam Integer hanger_type_id,
+    public ResponseEntity updateHangerType(@RequestParam Long hanger_type_id,
                                               @RequestBody HangerTypeEntity hangerType) {
         try {
             return ResponseEntity.ok("Группа переименована: " + hangerTypeService.updateData(hanger_type_id, hangerType));
