+++ src/main/java/com/example/mainproject/repository/ActionOperationTypeRepo.java
@@ -3,5 +3,5 @@ package com.example.mainproject.repository;
 import com.example.mainproject.entity.ActionOperationTypeEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActionOperationTypeRepo extends CrudRepository<ActionOperationTypeEntity, Integer> {
+public interface ActionOperationTypeRepo extends CrudRepository<ActionOperationTypeEntity, Long> {
 }
