+++ src/main/java/com/example/mainproject/repository/EquipmentStateRepo.java
@@ -3,5 +3,5 @@ package com.example.mainproject.repository;
 import com.example.mainproject.entity.EquipmentStateEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EquipmentStateRepo extends CrudRepository<EquipmentStateEntity, Integer> {
+public interface EquipmentStateRepo extends CrudRepository<EquipmentStateEntity, Long> {
 }
