+++ src/main/java/com/example/mainproject/service/ActionService.java
@@ -26,7 +26,7 @@ public class ActionService {
         return (List<ActionEntity>) actionRepo.findAll();
     }
 
-    public ActionEntity createData(ActionEntity actionEntity, Integer idWell, Integer idActionType){
+    public ActionEntity createData(ActionEntity actionEntity, Integer idWell, Long idActionType){
         WellEntity wellEntity = wellRepo.findById(idWell).get();
         ActionTypeEntity actionType = actionTypeRepo.findById(idActionType).get();
         actionEntity.setWell_id(wellEntity);
